+++ src/rtlib/file_inputstr.c
@@ -27,7 +27,7 @@ FBCALL FBSTRING *fb_FileStrInput( ssize_t bytes, int fnum )
         if( FB_HANDLE_IS_SCREEN(handle) )
         {
             dst->data[0] = 0;
-            while( read_count!=bytes ) {
+            while( read_count!=(size_t)bytes ) {
                 res = fb_FileGetDataEx( handle,
                                         0,
                                         dst->data + read_count,
@@ -61,7 +61,7 @@ FBCALL FBSTRING *fb_FileStrInput( ssize_t bytes, int fnum )
         /* add the null-term */
         dst->data[read_count] = '\0';
 
-        if( read_count != bytes )
+        if( read_count != (size_t)bytes )
         {
             fb_hStrSetLength( dst, read_count );
         }
