+++ src/rtlib/file_put.c
@@ -19,7 +19,7 @@ int fb_FilePutDataEx
 		return fb_ErrorSetNum( FB_RTERROR_ILLEGALFUNCTIONCALL );
 
 	if( pos < 0 )
-		return fb_ErrorSetNum( FB_RTERROR_ILLEGALFUNCTIONCALL );		
+		return fb_ErrorSetNum( FB_RTERROR_ILLEGALFUNCTIONCALL );
 
     FB_LOCK();
 
@@ -62,7 +62,7 @@ int fb_FilePutDataEx
          * The device must also support the SEEK method and the length
          * must be non-null */
 
-		if( length != handle->len )
+		if( length != (size_t)handle->len )
 			res = fb_ErrorSetNum( FB_RTERROR_FILEIO );
 
         size_t skip_size = (handle->len -
